Make server rendering possible
Adds the `exenv` module to determine whether the DOM can be accessed at
runtime, preventing them from being accessed if they are not.

When rendering on the server, window measurements provided by the screen
helper will fall back to the existing defaults. node-waves will not be
required on the server and a mock object will be exported instead
(though in theory none of its methods should ever be called from the server).